Adds support for the supermicro ivkm (that I have)
This now bastardizes the drac-kvm name by also supporting the supermicro
style ikvm (of which I have one). The version for the ivkm is poorly
identified as #1 for now (it should be changed to have version &
vendor). The ivkm is detected by attempting to post to the ikvm
login.cgi